Fixed clang compiler error for parameter_deduction types
