+++ include/simppl/detail/parameter_deduction.h
@@ -16,6 +16,33 @@ namespace simppl
 namespace dbus
 {
 
+struct oneway
+{
+   typedef oneway real_type;
+};
+
+
+// marker
+template<typename T>
+struct in
+{
+   typedef T real_type;
+};
+
+// marker
+template<typename T>
+struct out
+{
+   typedef T real_type;
+};
+
+// marker
+template<typename T>
+struct _throw
+{
+   typedef T real_type;
+};
+
 namespace detail
 {
